+++ epilepsy12/common_view_functions/recalculate_form_generate_response.py
@@ -350,9 +350,9 @@ def avoid_fields(model_instance):
         return META_VARIABLES + [
             "registration",
             "multiaxial_diagnosis",
-            "episode",
+            "episodes",
             "syndrome",
-            "comorbidity",
+            "comorbidities",
         ]
 
     elif model_class_name == "Management":
