+++ epilepsy12/serializers/keyword_serializer.py
@@ -1,0 +1,15 @@
+# Python imports
+
+# Django imports
+
+# DRF imports
+from rest_framework import serializers
+
+# E12 imports
+from epilepsy12.models import Keyword
+
+
+class KeywordSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Keyword
+        fields = ["keyword", "category"]
